+++ src/routes/Routes.tsx
@@ -1,5 +1,5 @@
+import { SavingsCalculatorPage } from 'domain/savingsCalculator/SavingsCalculatorPage';
 import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom';
-import { SavingsCalculatorPage } from './SavingsCalculatorPage';
 
 const router = createBrowserRouter([
   {
